#90 move oauth helper class to own file
